Remove inner lexer module
Before:

neoqo : crate
 └── lexer : private
     └── lexer : private

After:

neoqo : crate
 └── lexer : private
